requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
fastapi==0.115.8
frozenlist==1.5.0
gevent==24.11.1
greenlet==3.1.1
h11==0.14.0
httptools==0.6.4
idna==3.10
multidict==6.1.0
Pillow>=9.5.0
propcache==0.3.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyinstaller>=6.5.0
pystray>=0.19.4
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
setuptools==75.8.0
sniffio==1.3.1
SQLAlchemy==2.0.38
sse-starlette==2.2.1
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
websocket==0.2.1
websocket-client==1.8.0
websockets==15.0
yarl==1.18.3
zope.event==5.0
zope.interface==7.2
